+++ demisto_sdk/commands/validate/default_config.toml
@@ -67,7 +67,6 @@ select = [
     "RM104", "RM105", "RM113", "RM114",
     "CL100",
     "GF100", "GF101",
-    "GR104",
     "IF100", "IF101", "IF102", "IF103", "IF104", "IF105", "IF106",
     "IM101", "IM108",
     "RP101"
