+++ demisto_sdk/commands/validate/validators/SC_validators/SC109_script_name_is_not_unique_validator_list_files.py
@@ -3,20 +3,20 @@ from __future__ import annotations
 from typing import Iterable, List
 
 from demisto_sdk.commands.common.constants import ExecutionMode
-from demisto_sdk.commands.content_graph.objects.pack import Pack
+from demisto_sdk.commands.content_graph.objects.script import Script
 from demisto_sdk.commands.validate.validators.base_validator import (
     BaseValidator,
     ValidationResult,
 )
-from demisto_sdk.commands.validate.validators.GR_validators.GR104_is_pack_display_name_already_exists import (
-    IsPackDisplayNameAlreadyExistsValidator,
+from demisto_sdk.commands.validate.validators.SC_validators.SC109_script_name_is_not_unique_validator import (
+    DuplicatedScriptNameValidator,
 )
 
-ContentTypes = Pack
+ContentTypes = Script
 
 
-class IsPackDisplayNameAlreadyExistsValidatorListFiles(
-    IsPackDisplayNameAlreadyExistsValidator, BaseValidator[ContentTypes]
+class DuplicatedScriptNameValidatorListFiles(
+    DuplicatedScriptNameValidator, BaseValidator[ContentTypes]
 ):
     expected_execution_mode = [ExecutionMode.SPECIFIC_FILES, ExecutionMode.USE_GIT]
